% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_from_text.R
\name{infer}
\alias{infer}
\title{Infer without target. \n
Get the sentence Sentiment without an target mention.
This function is provided for convenience but the model was not
trained for this.}
\usage{
infer(text = NULL)
}
\arguments{
\item{text}{the sentence}
}
\value{
the sentiment
}
\description{
Infer without target. \n
Get the sentence Sentiment without an target mention.
This function is provided for convenience but the model was not
trained for this.
}
\examples{
infer("I like Peter")
}
