% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\docType{data}
\name{pack}
\alias{pack}
\title{Package Setup}
\format{
An object of class \code{environment} of length 0.
}
\usage{
pack
}
\arguments{
\item{pkgname}{}
}
\value{

}
\description{
Package Setup
}
\examples{


#
}
\keyword{datasets}
