% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_from_text.R
\name{infer}
\alias{infer}
\title{Title infer_from_text}
\usage{
infer(left, target, right)
}
\arguments{
\item{left}{the text from the beginning of the sentence to the mention of the
target (will be empty if the sentence starts with the mention of the target).
Make sure to include a space if there is one between the left phrase and the
target mention, e.g., "I like Peter." -> left="I like ", target="Peter", right=".".""}

\item{target}{The mention of the target.}

\item{right}{The text from the end of the target to the end of the sentence.
Make sure to include a space if there is one between the target and the next word
of the sentence.}
}
\value{

}
\description{
Title infer_from_text
}
\examples{


}
