% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_from_text.R
\name{infer_by_index}
\alias{infer_by_index}
\title{Infer by Index}
\usage{
infer_by_index(
  text = NULL,
  target_mention_from = NULL,
  target_mention_to = NULL
)
}
\arguments{
\item{text}{: The whole text.}

\item{target_mention_from}{Beginning Index of Target.}

\item{target_mention_to}{End Index of Target.}
}
\value{
the sentiment
}
\description{
Infer by Index
}
\examples{
print(sentiment[0])
}
